+++ packages/modernjs-v3/rslib.config.ts
@@ -1,0 +1,60 @@
+import { defineConfig } from '@rslib/core';
+import { pluginReact } from '@rsbuild/plugin-react';
+
+const sharedLibOptions = {
+  bundle: false,
+  externalHelpers: true,
+  outBase: 'src',
+} as const;
+
+export default defineConfig({
+  source: {
+    entry: {
+      index: ['./src/**/*.{ts,tsx,js,jsx}', '!./src/**/*.spec.*'],
+    },
+  },
+  plugins: [
+    pluginReact({
+      swcReactOptions: {
+        runtime: 'automatic',
+      },
+    }),
+  ],
+  lib: [
+    {
+      ...sharedLibOptions,
+      format: 'cjs',
+      syntax: 'es2019',
+      dts: false,
+      output: {
+        distPath: {
+          root: './dist/cjs',
+        },
+      },
+    },
+    {
+      ...sharedLibOptions,
+      format: 'esm',
+      syntax: 'es5',
+      dts: false,
+      output: {
+        distPath: {
+          root: './dist/esm',
+        },
+      },
+    },
+    {
+      ...sharedLibOptions,
+      format: 'esm',
+      syntax: 'es2019',
+      dts: {
+        distPath: './dist/types',
+      },
+      output: {
+        distPath: {
+          root: './dist/esm-node',
+        },
+      },
+    },
+  ],
+});